fix: Drop pipefail from the interpreter behavior optimization logic
The pipefail shell intepreter behavior is error-prone and will trigger
unexpected premature program termination if the output of the previous
pipeline component unexpectedly exceeds the size of the pipe buffer,
triggering the SIGPIPE signal and thus causing that component to return
exit status 141, which results in the entire pipeline failed due to
`pipefail`, interrupting the entire program execution due to `errexit`.

Developers who know how to deal with this and wishes to increase the
checking should set the behavior themselves, not by the template.

Fixes #2.

Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>